支持微信好友使用echo功能
